feat(dicom): implement DatabaseHandler and DICOMIndexer for database operations and indexing
