+++ .gitignore
@@ -221,4 +221,4 @@ data/**
 .imgtools/**/*.log
 
 *.dcm
-dicom_index.sqlite
+**/*.sqlite
