+++ src/imgtools/dicom/index/database/__init__.py
@@ -1,0 +1,3 @@
+from .database_handler import DatabaseHandler
+
+__all__ = ['DatabaseHandler']
